scrapy/requirements.txt
requests>=2.26.0
beautifulsoup4>=4.10.0
Pillow>=9.0.0
cairosvg>=2.5.0